MatN needed updates and additional information

+++ cinelerra-5.1/blds/bld_prepare.sh
@@ -55,7 +55,7 @@ case "$dir" in
     ctags patch gcc-c++ perl-XML-XPath libtiff-devel python dvdauthor \
     gettext-devel inkscape udftools autoconf automake numactl-devel \
     jbigkit-devel libvdpau-devel libva-devel gtk2-devel mesa-vdpau-drivers \
-    pulseaudio-libs-devel libtool
+    pulseaudio-libs-devel libtool patchelf boost-devel
   ;;
 "suse" | "leap" | "tumbleweed")
   zypper -n install nasm gcc gcc-c++ zlib-devel texinfo libpng16-devel \
@@ -67,8 +67,9 @@ case "$dir" in
     ilmbase-devel fftw3-devel libsndfile-devel libtheora-devel flac-devel \
     libtiff-devel inkscape cmake patch libnuma-devel lzma-devel udftools git \
     yasm autoconf automake rpm-build libjbig-devel libvdpau-devel libva-devel \
-    gtk2-devel libusb-1_0-devel libpulse-devel libtool python
-    if [ ! -f /usr/lib64/libtermcap.so ]; then
+    gtk2-devel libusb-1_0-devel libpulse-devel libtool python patchelf \
+    libboost_regex-devel libboost_filesystem-devel
+    if [ ! -f /usr/lib64/libtermcap.so && ! -L /usr/lib64/libtermcap.so ]; then
       ln -s libtermcap.so.2 /usr/lib64/libtermcap.so
     fi
   ;;
@@ -81,7 +82,8 @@ case "$dir" in
     fonts-dejavu libopenexr-dev festival libfftw3-dev gdb libusb-1.0-0-dev \
     libdc1394-22-dev libflac-dev libjbig-dev libvdpau-dev libva-dev \
     inkscape libsndfile1-dev libtheora-dev cmake udftools libxml2-utils git \
-    autoconf automake debhelper libgtk2.0-dev libpulse-dev python
+    autoconf automake debhelper libgtk2.0-dev libpulse-dev python \
+    patchelf libboost-filesystem-dev libboost-regex-dev fuse
   ;;
 #"ub16-10")
 #  apt-get -y install libx264-dev libx265-dev libvpx-dev libmjpegtools-dev
@@ -94,7 +96,8 @@ case "$dir" in
     libdc1394-22-dev libiec61883-dev libflac-dev libjbig-dev libusb-1.0-0-dev \
     libvdpau-dev libva-dev libsndfile1-dev libtheora-dev cmake udftools \
     libxml2-utils git inkscape autoconf automake debhelper libgtk2.0-dev \
-    libpulse-dev libtool python
+    libpulse-dev libtool python \
+    patchelf libboost-filesystem-dev libboost-regex-dev
   ;;
  *)
   echo "unknown os: $dir"
